[DEV-3647] More transaction states (#1974)
* [DEV-3647] Added more transaction states

* [DEV-3647] Renaming
